Rename serde_util module to key_value_pair
