+++ src/util/mod.rs
@@ -1,7 +1,7 @@
 //! Module containing helpers and util functions that are not specific to any DSL
 
-mod serde_util;
+mod key_value_pair;
 mod should_skip;
 
-pub(crate) use self::serde_util::*;
+pub(crate) use self::key_value_pair::*;
 pub(crate) use self::should_skip::*;
